Generic Template UI generator.
A form gen application that parses a json data structure to create a UI.  The type of widget used for a given UI element is determined by the type of a given keys values, i.e. "Author" : "Cameron" would create a text box with "Author" followed by an editable text box with "Cameron" to its right.